backend/auth_service/requirements.txt
fastapi
uvicorn
psycopg2
sqlalchemy
python-dotenv
passlib[bcrypt]
pyjwt
python-jose
python-multipart